feature: 感谢aaronuu提交的代码合并增加统一验证模块
