--- mate-core/mate-starter-validator/src/main/java/vip/mate/core/validators/StatusValue.java
@@ -1,42 +1,0 @@
-package vip.mate.core.validators;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import javax.validation.Constraint;
-import javax.validation.Payload;
-
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-/**
- * 状态校验，校验参数的状态是否是 StatusEnum 中的值
- *
- * @author aaronuu
- */
-@Documented
-@Constraint(validatedBy = StatusValueValidator.class)
-@Target({ElementType.FIELD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface StatusValue {
-
-  String message() default "不正确的状态标识";
-
-  Class[] groups() default {};
-
-  Class<? extends Payload>[] payload() default {};
-
-  /**
-   * 是否必填
-   * <p>
-   * 如果必填，在校验的时候本字段没值就会报错
-   */
-  boolean required() default true;
-
-  @Target({ElementType.FIELD, ElementType.PARAMETER})
-  @Retention(RUNTIME)
-  @Documented @interface List {
-    StatusValue[] value();
-  }
-}
