+++ mate-core/pom.xml
@@ -69,7 +69,7 @@
         <module>mate-starter-rabbit</module>
         <module>mate-starter-mybatis</module>
         <module>mate-starter-idempotent</module>
-      <module>mate-starter-validator</module>
+        <module>mate-starter-validator</module>
     </modules>
 
     <properties>
@@ -123,7 +123,7 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-<!--            <scope>provided</scope>-->
+            <!--            <scope>provided</scope>-->
         </dependency>
     </dependencies>
 
@@ -257,9 +257,9 @@
                 <sentinel.port>8719</sentinel.port>
                 <sentinel.dashboard>127.0.0.1:8080</sentinel.dashboard>
             </properties>
-<!--            <activation>-->
-<!--                <activeByDefault>true</activeByDefault>-->
-<!--            </activation>-->
+            <!--            <activation>-->
+            <!--                <activeByDefault>true</activeByDefault>-->
+            <!--            </activation>-->
         </profile>
         <profile>
             <id>test</id>
@@ -296,9 +296,9 @@
                 <sentinel.port>8719</sentinel.port>
                 <sentinel.dashboard>172.30.0.80:8088</sentinel.dashboard>
             </properties>
-<!--            <activation>-->
-<!--                <activeByDefault>true</activeByDefault>-->
-<!--            </activation>-->
+            <!--            <activation>-->
+            <!--                <activeByDefault>true</activeByDefault>-->
+            <!--            </activation>-->
         </profile>
     </profiles>
